Add basis levels
For wavelet bases the truncation should occur based upon the scale.
This commit adds a "level" attribute to the basis evaluations which
control the order in which the basis functions are truncated. For
Fourier or cosine this is simply 1:n, for wavelets this is the
sequence of the scale index.